Added advanced input group to the shiny apps This advanced group has settings for a custom site_dir for shiny apps Added migration file for the new field to shiny apps instances Moved SubdomainInputGroup to separate module Added missing closing div to the create_base.html
