+++ templates/apps/create_base.html
@@ -23,6 +23,7 @@
             {% endif %}
 
             </div>
+            <div>
                 {% crispy form %}
             </div>
         </div>
